+++ src/components/ui/switch.tsx
@@ -27,7 +27,7 @@ const Switch = forwardRef<HTMLInputElement, SwitchProps>(
 
 				<div
 					className={cn(
-						'h-6 w-11 rounded-full bg-zinc-600 peer-checked:bg-teal-500 peer-focus:outline-none peer-focus:ring-2 peer-focus:ring-teal-400 peer-focus:ring-offset-2 peer-focus:ring-offset-zinc-900 transition-colors',
+						'h-6 w-11 rounded-full bg-zinc-600 peer-checked:bg-primary-500 peer-focus:outline-none peer-focus:ring-2 peer-focus:ring-primary-400 peer-focus:ring-offset-2 peer-focus:ring-offset-zinc-900 transition-colors',
 						"after:absolute after:top-0.5 after:left-[2px] after:h-5 after:w-5 after:rounded-full after:border after:border-zinc-700 after:bg-white after:transition-all after:content-[''] peer-checked:after:translate-x-full peer-checked:after:border-white",
 						className
 					)}